+++ src/main/scala/com/jcdecaux/datacorp/spark/storage/v2/connector/ExcelConnector.scala
@@ -1,6 +1,8 @@
 package com.jcdecaux.datacorp.spark.storage.v2.connector
 
 import com.jcdecaux.datacorp.spark.internal.Logging
+import com.jcdecaux.datacorp.spark.util.ConfigUtils
+import com.typesafe.config.Config
 import org.apache.spark.sql._
 import org.apache.spark.sql.types.StructType
 
@@ -36,10 +38,48 @@ class ExcelConnector(val spark: SparkSession,
                      var workbookPassword: Option[String] = None,
                      var schema: Option[StructType] = None,
                      var saveMode: SaveMode = SaveMode.Overwrite
-                    ) extends Connector[DataFrame] with Logging {
+                    ) extends Connector with Logging {
 
   var reader: DataFrameReader = _
   var writer: DataFrameWriter[Row] = _
+
+  def this(spark: SparkSession, config: Config) = this(
+    spark = spark,
+    path =
+      ConfigUtils.getAs[String](config, "path").get,
+    useHeader =
+      ConfigUtils.getAs[String](config, "useHeader").get,
+    dataAddress =
+      ConfigUtils.getAs[String](config, "dataAddress").getOrElse("A1"),
+    treatEmptyValuesAsNulls =
+      ConfigUtils.getAs[String](config, "treatEmptyValuesAsNulls").getOrElse("true"),
+    inferSchema =
+      ConfigUtils.getAs[String](config, "inferSchema").getOrElse("false"),
+    addColorColumns =
+      ConfigUtils.getAs[String](config, "addColorColumns").getOrElse("false"),
+    timestampFormat =
+      ConfigUtils.getAs[String](config, "timestampFormat").getOrElse("yyyy-mm-dd hh:mm:ss.000"),
+    dateFormat =
+      ConfigUtils.getAs[String](config, "dateFormat").getOrElse("yyyy-mm-dd"),
+    maxRowsInMemory =
+      ConfigUtils.getAs[Long](config, "maxRowsInMemory"),
+    excerptSize =
+      ConfigUtils.getAs[Long](config, "excerptSize").getOrElse(10),
+    workbookPassword =
+      ConfigUtils.getAs[String](config, "workbookPassword"),
+    schema =
+      if (ConfigUtils.getAs[String](config, "schema").isDefined) {
+        Option(StructType.fromDDL(ConfigUtils.getAs[String](config, "schema").get))
+      } else {
+        None
+      },
+    saveMode =
+      if (ConfigUtils.getAs[String](config, "saveMode").isDefined) {
+        SaveMode.valueOf(ConfigUtils.getAs[String](config, "saveMode").get)
+      } else {
+        SaveMode.Overwrite
+      }
+  )
 
   override def read(): DataFrame = {
     reader = spark
@@ -70,7 +110,9 @@ class ExcelConnector(val spark: SparkSession,
       .option("dateFormat", dateFormat) // Optional, default: yy-m-d h:mm
 
     saveMode match {
-      case SaveMode.Append => writer.mode("append")
+      case SaveMode.Append =>
+        log.warn("It seems that the Append save mode doesn't work properly. Please make sure that your data are written correctly")
+        writer.mode("append")
       case SaveMode.Overwrite => writer.mode("overwrite")
       case _ => throw new IllegalArgumentException(s"Unknown save mode: $saveMode")
     }