+++ src/test/scala/com/jcdecaux/datacorp/spark/storage/v2/connector/CSVConnectorSuite.scala
@@ -2,6 +2,7 @@ package com.jcdecaux.datacorp.spark.storage.v2.connector
 
 import java.io.File
 
+import com.jcdecaux.datacorp.spark.config.Properties
 import com.jcdecaux.datacorp.spark.storage.SparkRepositorySuite
 import com.jcdecaux.datacorp.spark.{SparkSessionBuilder, TestObject}
 import org.apache.spark.sql.{Dataset, SaveMode, SparkSession}
@@ -16,8 +17,7 @@ class CSVConnectorSuite extends FunSuite {
 
   import SparkRepositorySuite.deleteRecursively
 
-  test("IO") {
-
+  test("IO CSVConnector") {
     import spark.implicits._
     val testTable: Dataset[TestObject] = Seq(
       TestObject(1, "p1", "c1", 1L),
@@ -35,5 +35,25 @@ class CSVConnectorSuite extends FunSuite {
     assert(df.count() === 6)
     deleteRecursively(new File(path))
 
+  }
+
+  test(s"IO with auxiliary CSVConnector constructor") {
+    import spark.implicits._
+
+    val testTable: Dataset[TestObject] = Seq(
+      TestObject(1, "p1", "c1", 1L),
+      TestObject(2, "p2", "c2", 2L),
+      TestObject(3, "p3", "c3", 3L)
+    ).toDS()
+
+    val connector = new CSVConnector(spark, Properties.csvConfig)
+
+    connector.write(testTable.toDF())
+    connector.write(testTable.toDF())
+
+    val df = connector.read()
+    df.show()
+    assert(df.count() === 6)
+    deleteRecursively(new File(Properties.csvConfig.getString("path")))
   }
 }