+++ src/test/scala/com/jcdecaux/datacorp/spark/storage/v2/connector/ExcelConnectorSuite.scala
@@ -3,10 +3,11 @@ package com.jcdecaux.datacorp.spark.storage.v2.connector
 import java.io.File
 import java.sql.{Date, Timestamp}
 
+import com.jcdecaux.datacorp.spark.config.Properties
 import com.jcdecaux.datacorp.spark.storage.SparkRepositorySuite
-import com.jcdecaux.datacorp.spark.{SparkSessionBuilder, TestObject2}
+import com.jcdecaux.datacorp.spark.{SparkSessionBuilder, TestObject, TestObject2}
 import org.apache.spark.sql.types._
-import org.apache.spark.sql.{Dataset, SparkSession}
+import org.apache.spark.sql.{Dataset, SaveMode, SparkSession}
 import org.scalatest.FunSuite
 
 class ExcelConnectorSuite extends FunSuite {
@@ -24,7 +25,8 @@ class ExcelConnectorSuite extends FunSuite {
     TestObject2("string3", 5, 0.000000001685400132103450D, new Timestamp(1557153268000L), new Date(1557100800000L), 999999999999999999L)
   ).toDS()
 
-  test("IO with default parameters") {
+  test("IO with default excel connector parameters") {
+
     val excelConnector = new ExcelConnector(spark, path, "true")
 
     testTable.toDF.show(false)
@@ -41,7 +43,8 @@ class ExcelConnectorSuite extends FunSuite {
 
   }
 
-  test("IO with customized format") {
+  test("IO with customized format excel connector") {
+
     val schema: StructType = StructType(Array(
       StructField("col1", StringType),
       StructField("col2", IntegerType),
@@ -57,7 +60,8 @@ class ExcelConnectorSuite extends FunSuite {
       useHeader = "true",
       timestampFormat = "dd/mm/yyyy hh:mm:ss",
       dateFormat = "dd/mm/yy",
-      schema = Some(schema)
+      schema = Some(schema),
+      saveMode = SaveMode.Append // TODO SaveMode.Append seems not working.
     )
 
     excelConnector.write(testTable.toDF)
@@ -71,6 +75,28 @@ class ExcelConnectorSuite extends FunSuite {
     assert(df.head.getAs[Date]("col5").getTime === 1557100800000L)
 
     deleteRecursively(new File(path))
+
+  }
+
+  test("IO with excel connector auxiliary constructor") {
+
+    import spark.implicits._
+    val testTable: Dataset[TestObject] = Seq(
+      TestObject(1, "p1", "c1", 1L),
+      TestObject(2, "p2", "c2", 2L),
+      TestObject(3, "p3", "c3", 3L)
+    ).toDS()
+
+    val connector = new ExcelConnector(spark = spark, config = Properties.excelConfig)
+
+    testTable.toDF.show()
+    connector.write(testTable.toDF)
+
+    val df = connector.read()
+
+    df.show()
+    assert(df.count() === 3)
+    deleteRecursively(new File("src/test/resources/test_config_excel.xlsx"))
 
   }